+++ ESP-WROVER-KIT/src/Led_Control.cpp
@@ -1,17 +1,19 @@
 #include <Arduino.h>
+#include <LED_Control.h>
 
 const uint8_t LED_PIN = 4;                          // PWM Pin
 const double LED_FREQ = 5000;                       // PWM Frequency
 const uint8_t LED_CH = 0;                           // PWM Channel
 const uint8_t LED_RES = 8;                          // 8-bit ADC Resolution
 
-uint8_t dutyCycle = 255;                       // 8-bit Duty Cycle
-bool light = true;
+uint8_t dutyCycle = 163;                            // 8-bit Duty Cycle
+bool light;
 
 bool setupLED(){
     ledcSetup(LED_CH, LED_FREQ, LED_RES);
     ledcAttachPin(LED_PIN, LED_CH);
     ledcWrite(LED_CH, dutyCycle);
+    turnOnLight();
 }
 
 bool turnOnLight(){
@@ -36,5 +38,9 @@ void setLightIntensity(uint8_t intensity){
 }
 
 bool lightIsOn(){
+    ledcWrite(LED_CH, dutyCycle);
+    Serial.print("Light On: ");
+    Serial.print(light);
+    Serial.println("$");
     return light;
 }
