+++ ESP-WROVER-KIT/src/main.cpp
@@ -9,15 +9,14 @@
 #include <Adafruit_AS7341.h>
 #include <string>
 
+#include <Led_Control.h>
+#include <Serial_Processing.h>
+
 ///////////////////   CONSTANTS    ///////////////
 
-const uint8_t LED_PIN = 4;       // PWM Pin
-const double LED_FREQ = 5000;       // PWM Frequency
-const uint8_t LED_CH = 0;         // PWM Channel
-const uint8_t LED_RES = 8;       // ADC Resolution
-const uint32_t LED_DUTY_CYCLE = 255;
+const String SMARTCLAMP_VERSION = "0.15";
 
-///////////////////   OBJECT DECLARATION    ///////////////
+///////////////////   GLOBAL VARIABLES    ///////////////
 
 Adafruit_AS7341 as7341;
 
@@ -48,9 +47,8 @@ void setup() {
   as7341.setGain(AS7341_GAIN_256X);
 
   // Setup LED PWM Signal.
-  ledcSetup(LED_CH, LED_FREQ, LED_RES);
-  ledcAttachPin(LED_PIN, LED_CH);
-  ledcWrite(LED_CH, LED_DUTY_CYCLE);
+  setupLED();
+ 
 
   Serial.println("Done with setup");
 }
